Fixed breaking prompt on older zsh versions, and added mercurial support.
On zsh 4.3.6, colors in the PS1 cannot be written out like %F{white},
but require escape sequences. Since I would prefer to keep my PS1 as
readable as possible, I have left it the way it was, and default to
"prompt adam1" on older zsh installations.

I also added mercurial support for listing the current branch.